pyproject.toml
[tool.poetry]
name = "midori"
version = "0.1.0"
description = "Spotify API wrapper."
authors = ["PureFunctor <purefunctor@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
attrs = "^21.2.0"
cattrs = "^1.7.1"
httpx = "^0.18.2"

[tool.poetry.dev-dependencies]
Sphinx = "^4.0.2"
black = "^21.6b0"
coverage = {extras = ["toml"], version = "^5.5"}
flake8 = "^3.9.2"
flake8-annotations = "^2.6.2"
flake8-docstrings = "^1.6.0"
flake8-import-order = "^0.18.1"
mypy = "^0.902"
pre-commit = "^2.13.0"
pytest = "^6.2.4"
pytest-mock = "^3.6.1"
sphinx-autobuild = "^2021.3.14"
sphinx-rtd-theme = "^0.5.2"

[tool.coverage.run]
branch = true
source = ["midori"]
omit = ["*tests*"]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
